+++ src/datumaro/plugins/transforms.py
@@ -22,6 +22,8 @@ import pycocotools.mask as mask_utils
 from pandas.api.types import CategoricalDtype
 
 import datumaro.util.mask_tools as mask_tools
+from datumaro.components.algorithms.hash_key_inference.explorer import Explorer
+from datumaro.components.algorithms.hash_key_inference.hashkey_util import calculate_hamming
 from datumaro.components.annotation import (
     AnnotationType,
     Bbox,
@@ -40,6 +42,7 @@ from datumaro.components.annotation import (
     TabularCategories,
 )
 from datumaro.components.cli_plugin import CliPlugin
+from datumaro.components.dataset import Dataset
 from datumaro.components.dataset_base import DEFAULT_SUBSET_NAME, DatasetInfo, DatasetItem, IDataset
 from datumaro.components.errors import (
     AnnotationTypeError,
@@ -1348,9 +1351,21 @@ class RemoveAttributes(ItemTransform):
 
 class Correct(Transform, CliPlugin):
     """
-    Correct the dataset from a validation report.
-    A user can should feed into validation_reports.json from validator to correct the dataset.
-    This helps to refine the dataset by rejecting undefined labels, missing annotations, and outliers.
+    This class provides functionality to correct and refine a dataset based on a validation report.|n
+    It processes a validation report (typically in JSON format) to identify and rectify various |n
+    dataset issues, such as undefined labels, missing annotations, outliers, empty labels/captions,|n
+    and unnecessary characters in captions. The correction process includes:|n
+    |n
+    - Adding missing labels and attributes.|n
+    - Removing or adjusting annotations with invalid or anomalous values.|n
+    - Filling in missing labels and captions with appropriate values.|n
+    - Removing unnecessary characters from text-based annotations like captions.|n
+    - Handling outliers by capping values within specified bounds.|n
+    - Updating dataset categories and annotations according to the corrections.|n
+    |n
+    The class is designed to be used as part of a command-line interface (CLI) and can be |n
+    configured with different validation reports. It integrates with the dataset extraction |n
+    process, ensuring that corrections are applied consistently across the dataset.|n
     """
 
     @classmethod
@@ -1746,13 +1761,15 @@ class Correct(Transform, CliPlugin):
 
 class AstypeAnnotations(ItemTransform):
     """
-    Enables the conversion of annotation types for the categories and individual items within a dataset.|n
+    Converts the types of annotations within a dataset based on a specified mapping.|n
     |n
-    Based on a specified mapping, it transforms the annotation types,|n
-    changing them to 'Label' if they are categorical, and to 'Caption' if they are of type string, float, or integer.|n
+    This transform changes annotations to 'Label' if they are categorical, and to 'Caption'
+    if they are of type string, float, or integer. This is particularly useful when working
+    with tabular data that needs to be converted into a format suitable for specific machine
+    learning tasks.|n
     |n
     Examples:|n
-        - Convert type of `title` annotation|n
+        - Converts the type of a `title` annotation:|n
 
         .. code-block::
 
@@ -2004,3 +2021,64 @@ class Clean(ItemTransform):
             refined_annotations.append(ann)
 
         return self.wrap_item(item, media=refined_media, annotations=refined_annotations)
+
+
+class PseudoLabeling(ItemTransform):
+    """
+    A class used to assign pseudo-labels to items in a dataset based on
+    their similarity to predefined labels.|n
+    |n
+    This class leverages hashing techniques to compute the similarity
+    between dataset items and a set of predefined labels.|n
+    It assigns the most similar label as a pseudo-label to each item.
+    This is particularly useful in semi-supervised
+    learning scenarios where some labels are missing or uncertain.|n
+    |n
+    Attributes:|n
+        - extractor : IDataset|n
+        The dataset extractor that provides access to dataset items and their annotations.|n
+        - labels : Optional[List[str]]|n
+        A list of label names to be used for pseudo-labeling.
+        If not provided, all available labels in the dataset will be used.|n
+        - explorer : Optional[Explorer]|n
+        An optional Explorer object used to compute hash keys for items and labels.
+        If not provided, a new Explorer will be created.|n
+    """
+
+    def __init__(
+        self,
+        extractor: IDataset,
+        labels: Optional[List[str]] = None,
+        explorer: Optional[Explorer] = None,
+    ):
+        super().__init__(extractor)
+
+        self._categories = self._extractor.categories()
+        self._labels = labels
+        self._explorer = explorer
+        self._label_indices = self._categories[AnnotationType.label]._indices
+
+        if not self._labels:
+            self._labels = list(self._label_indices.keys())
+        if not self._explorer:
+            self._explorer = Explorer(Dataset.from_iterable(list(self._extractor)))
+
+        label_hashkeys = [
+            np.unpackbits(self._explorer._get_hash_key_from_text_query(label).hash_key, axis=-1)
+            for label in self._labels
+        ]
+        self._label_hashkeys = np.stack(label_hashkeys, axis=0)
+
+    def categories(self):
+        return self._categories
+
+    def transform_item(self, item: DatasetItem):
+        hashkey_ = np.unpackbits(self._explorer._get_hash_key_from_item_query(item).hash_key)
+        logits = calculate_hamming(hashkey_, self._label_hashkeys)
+        inverse_distances = 1.0 / (logits + 1e-6)
+        probs = inverse_distances / np.sum(inverse_distances)
+        ind = np.argsort(probs)[::-1]
+
+        pseudo = np.array(self._labels)[ind][0]
+        pseudo_annotation = [Label(label=self._label_indices[pseudo])]
+        return self.wrap_item(item, annotations=pseudo_annotation)